Move ClassDataStore.DisplayDate to AppStore
