+++ src/store/index.ts
@@ -3,6 +3,7 @@ import Vuex, { ActionContext } from 'vuex'
 import { Context } from '@nuxt/types/app'
 import { createProxy, extractVuexModule } from 'vuex-class-component'
 import jwtDecode from 'jwt-decode'
+import { AppStore } from '@/store/modules/app'
 import { UserStore } from '@/store/modules/user'
 import { ClassDataStore } from '@/store/modules/classData'
 
@@ -10,6 +11,7 @@ Vue.use(Vuex)
 
 export const store = new Vuex.Store({
   modules: {
+    ...extractVuexModule(AppStore),
     ...extractVuexModule(UserStore),
     ...extractVuexModule(ClassDataStore)
   }
@@ -29,6 +31,7 @@ export const actions = {
 }
 
 export const vxm = {
+  app: createProxy(store, AppStore),
   user: createProxy(store, UserStore),
   classData: createProxy(store, ClassDataStore)
 }
